refactor(parser): extract environment file handling into envFiles.ts
Move http-client.env.json parsing and the environment variable cascade
resolution (parseEnvironmentFile, ensureSharedEnvironment,
getEnvironmentNames, resolveEnvironmentVariables,
resolveEnvironmentVariablesWithSource) out of parser.ts into
src/lib/envFiles.ts. parser.ts keeps temporary re-exports.
